requirements.txt
# Paquetes principales

streamlit==1.43.2
langchain-community==0.3.20
transformers==4.50.0
huggingface-hub==0.29.3
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3